applications/backend/pyproject.toml
[build-system]
requires = ["setuptools>=70.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "omnibooker-backend"
version = "0.1.0"
description = "FastAPI backend for Omnibooker"
authors = [{ name = "Omnibooker" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.29",
  "psycopg[binary]>=3.2.1",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "httpx>=0.27.0",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "email-validator>=2.1.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "pytest-asyncio>=0.23.0"
]


